Add info parameter to Crypto
